Moved LDD code to commons project
