+++ src/main/java/gov/nasa/pds/harvest/HarvestServer.java
@@ -12,7 +12,6 @@ import org.eclipse.jetty.util.thread.QueuedThreadPool;
 
 import gov.nasa.pds.harvest.cfg.Configuration;
 import gov.nasa.pds.harvest.cfg.ConfigurationReader;
-import gov.nasa.pds.harvest.cfg.RegistryCfg;
 import gov.nasa.pds.harvest.dao.RegistryManager;
 import gov.nasa.pds.harvest.dao.SchemaUtils;
 import gov.nasa.pds.harvest.http.MemoryServlet;
@@ -20,6 +19,7 @@ import gov.nasa.pds.harvest.http.StatusServlet;
 import gov.nasa.pds.harvest.mq.MQClient;
 import gov.nasa.pds.harvest.mq.rmq.ConsumerFactory;
 import gov.nasa.pds.harvest.mq.rmq.RabbitMQClient;
+import gov.nasa.pds.registry.common.cfg.RegistryCfg;
 import gov.nasa.pds.registry.common.util.ExceptionUtils;
