+++ src/main/java/gov/nasa/pds/harvest/cfg/ConfigurationReader.java
@@ -7,6 +7,7 @@ import java.io.FileReader;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
+import gov.nasa.pds.registry.common.cfg.RegistryCfg;
 import gov.nasa.pds.registry.common.util.CloseUtils;
