+++ src/main/java/gov/nasa/pds/harvest/mq/ManagerCommandConsumer.java
@@ -3,9 +3,9 @@ package gov.nasa.pds.harvest.mq;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
-import gov.nasa.pds.harvest.cfg.RegistryCfg;
 import gov.nasa.pds.harvest.dao.RegistryManager;
 import gov.nasa.pds.harvest.mq.msg.ManagerMessage;
+import gov.nasa.pds.registry.common.cfg.RegistryCfg;
 import gov.nasa.pds.registry.common.es.service.ProductService;
 
 /**
