+++ src/test/java/gov/nasa/pds/TestInventoryConsumer.java
@@ -1,10 +1,10 @@
 package gov.nasa.pds;
 
 import gov.nasa.pds.harvest.cfg.HarvestCfg;
-import gov.nasa.pds.harvest.cfg.RegistryCfg;
 import gov.nasa.pds.harvest.mq.CollectionInventoryConsumer;
 import gov.nasa.pds.harvest.mq.msg.CollectionInventoryMessage;
 import gov.nasa.pds.harvest.util.Log4jConfigurator;
+import gov.nasa.pds.registry.common.cfg.RegistryCfg;
 
 
 public class TestInventoryConsumer
