+++ src/test/java/gov/nasa/pds/TestProductConsumer.java
@@ -4,12 +4,12 @@ import java.util.Arrays;
 import java.util.Set;
 
 import gov.nasa.pds.harvest.cfg.Configuration;
-import gov.nasa.pds.harvest.cfg.RegistryCfg;
 import gov.nasa.pds.harvest.dao.RegistryManager;
-import gov.nasa.pds.harvest.dao.SchemaDao;
 import gov.nasa.pds.harvest.mq.ProductConsumer;
 import gov.nasa.pds.harvest.mq.msg.ProductMessage;
 import gov.nasa.pds.harvest.util.Log4jConfigurator;
+import gov.nasa.pds.registry.common.cfg.RegistryCfg;
+import gov.nasa.pds.registry.common.es.dao.schema.SchemaDao;
 import gov.nasa.pds.registry.common.meta.FieldNameCache;
 
 
@@ -42,7 +42,7 @@ public class TestProductConsumer
 
         SchemaDao schemaDao = RegistryManager.getInstance().getSchemaDao();
         Set<String> fields = schemaDao.getFieldNames();
-        FieldNameCache.getInstance().set(fields);
+        FieldNameCache.getInstance().setSchemaFieldNames(fields);
     }
